All endpoints of categories found Delete and get of all endpoints found

+++ Recipes/src/Application/Command/Recipes/RemoveRecipe.php
@@ -2,8 +2,10 @@
 
 namespace Recipes\Application\Command\Recipes;
 
+use LIN3S\SharedKernel\Exception\Exception;
 use Recipes\Domain\Model\Recipes\RecipeId;
 use Recipes\Domain\Model\Recipes\RecipeRepository;
+use Recipes\Domain\Model\User\UserId;
 
 class RemoveRecipe
 {
@@ -16,6 +18,12 @@ class RemoveRecipe
 
     public function __invoke(RemoveRecipeCommand $command)
     {
+        $recipe = $this->repository->recipeOfId(RecipeId::generate($command->id()));
+
+        if($recipe === null || !$recipe->owner()->equals(UserId::generate($command->userId()))) {
+            throw new Exception('Invalid userId, is not your book');
+        }
+
         $this->repository->remove(RecipeId::generate($command->id()));
     }
 }
